kernel: move m4 interrupt string map to arch
The `ipsr_isr_number_to_str()` function maps interrupts to string names,
but is architecture specific so it shouldn't be in process.rs.
