+++ chips/tm4c129x/src/lib.rs
@@ -234,7 +234,7 @@ unsafe extern "C" fn hard_fault_handler() {
             ici_it,
             thumb_bit,
             exception_number,
-            kernel::process::ipsr_isr_number_to_str(exception_number),
+            cortexm4::ipsr_isr_number_to_str(exception_number),
             faulting_stack as u32,
             (_estack as *const ()) as u32,
             (&_ezero as *const u32) as u32,
